feat: Save last read position
Saves the reader's last position for each book and automatically restores it when the book is reopened.

The last read locator is now saved to SharedPreferences when the activity stops and retrieved when it starts. The calculation for determining the last read CFI in the web view has also been improved for both vertical and horizontal scrolling modes.